requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.2
charset-normalizer==2.1.1
click~=7.1.2
colorama==0.4.5
fastapi~=0.63.0
h11~=0.14.0
idna~=3.4
lxml==4.9.1
pydantic==1.10.2
requests==2.28.1
sniffio==1.3.0
soupsieve~=2.3.1
starlette~=0.13.6
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn~=0.13.4
xmltodict==0.12.0
xmltojson==2.0.1

pytest~=7.2.1
pip~=22.3
cryptography~=38.0.1
py~=1.11.0
attrs~=22.2.0
redis~=4.5.1
wcwidth~=0.2.5
pluggy~=1.0.0
exceptiongroup~=1.1.0
iniconfig~=2.0.0
numpy~=1.22.2
responses~=0.21.0
websockets~=10.3
spotipy~=2.22.1
six~=1.16.0
PyYAML~=6.0
httptools~=0.5.0
watchfiles~=0.17.0
decorator~=5.1.1
chardet~=3.0.4
itsdangerous~=2.1.2
python-multipart~=0.0.5
Jinja2~=3.0.3
setuptools~=49.2.1
python-dotenv~=0.21.0